chore: cutlass cleanup
